Add MUC07: Multi-user LLM-powered task list with zero code duplication
Extract AbstractTaskChatView base class that accepts signals via constructor,
enabling both single-user (UseCase18) and multi-user (MUC07) implementations
with zero UI code duplication.

Changes:
- Add ChatMessage signals to CollaborativeSignals for MUC07
- Add getLlmTasksSignal() and getLlmChatMessagesSignal() getters
- Add initializeSampleLLMTasks() initialization method
- Extract AbstractTaskChatView from UseCase18View
  - Accept signals via constructor parameters
  - Keep all UI logic (chat, grid, forms, statistics) unchanged
  - ~450 lines of reusable UI code
- Refactor UseCase18View to extend AbstractTaskChatView
  - Now only 42 lines (down from 485 lines)
  - Pass view-local signals via super()
  - Initialize sample tasks in constructor
- Create MUC07View for multi-user collaboration
  - ~70 lines total
  - Pass shared signals from CollaborativeSignals
  - Add ActiveUsersDisplay component
  - Per-user conversation IDs for independent LLM memory
  - All users see same task list and chat in real-time

Architecture:
- AbstractTaskChatView: Base class with all UI components and logic
- UseCase18View: Single-user with view-local signals
- MUC07View: Multi-user with shared signals from CollaborativeSignals
- Zero code duplication - only signal source differs

Net code increase: ~105 lines for full multi-user support
